Move src to main/java
